migrate cluster dns test to sig network
