+++ test/e2e/BUILD
@@ -39,7 +39,6 @@ go_library(
     srcs = [
         "e2e.go",
         "events.go",
-        "example_cluster_dns.go",
         "examples.go",
         "gke_local_ssd.go",
         "gke_node_pools.go",
